fix cloumn

+++ backend/schemas/schemas.py
@@ -19,7 +19,7 @@ class BaseOrders(BaseModel):
     customer_wallet: str
     deal_conditions: str
     token_address: str
-    token_amount: int
+    token_amount: float
     start_date: str
     end_date: str
 
@@ -31,7 +31,7 @@ class BaseOrders(BaseModel):
                     "customer_wallet": "wallet",
                     "deal_conditions": "conditions",
                     "token_address": "address",
-                    "token_amount": 100,
+                    "token_amount": 100.00,
                     "start_date": "01.01.2022 12:00",
                     "end_date": "02.01.2022 12:00",
                 }
@@ -53,7 +53,7 @@ class Orders(BaseOrders):
                     "customer_wallet": "wallet",
                     "deal_conditions": "conditions",
                     "token_address": "address",
-                    "token_amount": 100,
+                    "token_amount": 100.00,
                     "start_date": "2022-01-01T12:00:00",
                     "end_date": "2022-01-02T12:00:00",
                     "order_id": "77f796d4-71df-4147-af32-06e481fb850e",
